Separating everything into cpp and .h files. Tried to use forward definition but did not work for me.
