+++ Pokemon/Player.h
@@ -1,7 +1,10 @@
+#pragma once 
 #include <string>
 #include "Pokemon.h"
 
 using namespace std;
+
+//class Pokemon;
 
 class Player {
 public:
